Add minor cleanup and improvements
